Add client order id to query

+++ server_app/include/order/order.h
@@ -4,7 +4,7 @@
 #include <string>
 #include <json/json.h>
 
-using OrderId = uint64_t;
+using OrderId = size_t;
 
 class Order
 {
